Update patch set 2
Patch Set 2:

There are known to be devices that enabling this on will cause bad things to happen, apparently, so it should be opt-in.  I guess we could move it to common and provide an opt-out variable.  Then it just comes down to making sure we write 'enough' sysfs nodes for all devices in a single family.

Matt suggested doing it per device, so I did.

Patch-set: 2