Update patch set 1
Patch Set 1:

tbh, I'm not very fond of either suggestions. api.mk would be quite specific, for that I'd let it stay on android.mk (where we state the original android api level, and might state other android-related stuff in the future). And putting it on the device.mk, which at the moment only has the include of product/*mk, and putting it there would sort of defeat the purpose of the division...

Patch-set: 1