Update patch set 1
Patch Set 1:

OK, so...

Checking more into how this is actually used, it seems that:
* if SUPL_ES is disabled, a GPS Network Initiated notification will always be handled;
* if SUPL_ES is enabled, a GPS Network Initiated notification will only be handled if we're in emergency mode, otherwise it will be ignored.

This behavior doesn't seem to have changed, so it seems to me that setting SUPL_ES to 0 now only makes us reply to more NI notifications than previously. The case is different from what I thought in my previous comment (requests are always dealt with in emergency mode, no problem there), but I am still not sure that we want to change this, now because that would mean we would probably be giving GPS information in more cases than we need (possibly should) to.

As a next step, I think we should observe the changes on the HAL, to see when do we actually deal with NI requests, and if anything in those request cases changed to justify QC's default change from 1 to 0.

Patch-set: 1